swag/requirements.txt
Django
python-decouple
pymongo
python-dotenv
djangorestframework
requests
django-cors-headers
mongoengine
bcrypt